integrate step quiz action dispatcher in StepQuizReviewPresenter
