+++ app/src/main/java/org/stepik/android/presentation/base/dispatcher/ActionDispatcher.kt
@@ -1,0 +1,8 @@
+package org.stepik.android.presentation.base.dispatcher
+
+import io.reactivex.disposables.CompositeDisposable
+
+interface ActionDispatcher<Action, Message> {
+    fun handleAction(action: Action, onNewMessage: (Message) -> Unit)
+    val disposable: CompositeDisposable
+}
