+++ app/src/main/java/org/stepik/android/presentation/base/reducer/StateReducer.kt
@@ -1,4 +1,4 @@
-package org.stepik.android.presentation.step_quiz_review.reducer
+package org.stepik.android.presentation.base.reducer
 
 interface StateReducer<State, Message, Action> {
     fun reduce(state: State, message: Message): Pair<State, Set<Action>>
